Make reusable moderationLabel component
